feat(db): implement asynchronous database persistence layer
- Introduced DatabaseTask abstract interface for command-based DB operations.
- Implemented thread-safe SaveQueue with condition variables.
- Added DBWorker thread to decouple Game Logic from MySQL I/O.
- Wrapped PlayerSaver into SavePlayerTask for background execution.

+++ include/game/player/Currency.h
@@ -5,14 +5,14 @@
 class Currency
 {
 public:
-    bool Spend(int amount);
+    bool Spend(uint64_t amount);
 
-    void Add(int amount);
+    void Add(uint64_t amount);
 
     void Set(uint64_t value);
 
     uint64_t Get() const;
 
 private:
-    std::atomic<int> value_{0};
+    std::atomic<uint64_t> value_{0};
 };
